+++ catsup/compat.py
@@ -1,0 +1,9 @@
+import sys
+
+py = sys.version_info
+py3k = py >= (3, 0, 0)
+
+if py3k:
+    basestring = str
+    unicode = str
+