+++ catsup/deploy.py
@@ -23,12 +23,11 @@ def git(config):
     if os.path.exists(dot_git_path):
         if _call('git remote -v | grep %s' % config.deploy.git.repo) == 0:
             shutil.rmtree(dot_git_path)
-    if os.path.exists(dot_git_path):
-        _call('git init', silence=True)
+    if not os.path.exists(dot_git_path):
+        _call('git init')
         _call('git remote add origin %s' % config.deploy.git.repo)
-
-    _call('git branch -m %s' % config.deploy.git.branch, silence=True)
-    _call('git pull origin %s' % config.deploy.git.branch, silence=True)
+    if _call('git checkout %s' % config.deploy.git.branch) != 0:
+        _call('git branch -m %s' % config.deploy.git.branch)
     if config.deploy.git.delete:
         _call('rm -rf *')
 
@@ -37,12 +36,10 @@ def git(config):
     generator = Generator(config.path)
     generator.generate()
 
-    
-
     _call('git add .', silence=True)
     _call('git commit -m "Update at %s"' % str(datetime.datetime.utcnow()),
           silence=True)
-    _call('git push origin %s' % config.deploy.git.branch)
+    _call('git push origin %s --force' % config.deploy.git.branch)
 
 
 def rsync(config):